corrected betty

+++ backup/flush_buffer.c
@@ -7,7 +7,7 @@
  *
  * Return: number of printed characters
  */
-long int _flushbuff(char *buf, unsigned int *buf_i)
+long int _flushbuff(char *buf, long int *buf_i)
 {
 	long int nob = 0;
