+++ backup/format_handler.c
@@ -10,7 +10,7 @@
  * Return: the number of characters printed, -1 if format doesn't match
  */
 long int
-format_handler(va_list args, char format, char *buf, unsigned int *buf_i)
+format_handler(va_list args, char format, char *buf, long int *buf_i)
 {
 	long int b, nob = 0;
 	f_prt fmts[] = {{'c', print_c},
