+++ backup/print_str.c
@@ -8,7 +8,7 @@
  *
  * Return: number of characters written
  */
-long int print_s(va_list args, char *buf, unsigned int *buf_i)
+long int print_s(va_list args, char *buf, long int *buf_i)
 {
 	unsigned int idx, nob = 0;
 	char *str;
